+++ src/execution/createPipelineExecutor/00-createPipelineExecutor.ts
@@ -1,20 +1,22 @@
 import { spaceTrim } from 'spacetrim';
 import type { PartialDeep, Promisable, ReadonlyDeep } from 'type-fest';
-import { DEFAULT_CSV_SETTINGS } from '../../config';
-import { DEFAULT_INTERMEDIATE_FILES_STRATEGY } from '../../config';
-import { DEFAULT_IS_AUTO_INSTALLED } from '../../config';
-import { DEFAULT_IS_VERBOSE } from '../../config';
-import { DEFAULT_MAX_EXECUTION_ATTEMPTS } from '../../config';
-import { DEFAULT_MAX_PARALLEL_COUNT } from '../../config';
-import { DEFAULT_SCRAPE_CACHE_DIRNAME } from '../../config';
+import {
+    DEFAULT_CSV_SETTINGS,
+    DEFAULT_INTERMEDIATE_FILES_STRATEGY,
+    DEFAULT_IS_AUTO_INSTALLED,
+    DEFAULT_IS_VERBOSE,
+    DEFAULT_MAX_EXECUTION_ATTEMPTS,
+    DEFAULT_MAX_PARALLEL_COUNT,
+    DEFAULT_SCRAPE_CACHE_DIRNAME,
+} from '../../config';
 import { validatePipeline } from '../../conversion/validation/validatePipeline';
 import type { PipelineJson } from '../../pipeline/PipelineJson/PipelineJson';
 import { isPipelinePrepared } from '../../prepare/isPipelinePrepared';
 
-import { exportJson } from '../../utils/serialization/exportJson';
-import { serializeError } from '../../errors/utils/serializeError';
 import { assertsError } from '../../errors/assertsError';
+import { serializeError } from '../../errors/utils/serializeError';
 import type { InputParameters } from '../../types/typeAliases';
+import { exportJson } from '../../utils/serialization/exportJson';
 import type { ExecutionTask } from '../ExecutionTask';
 import { createTask } from '../ExecutionTask';
 import type { PipelineExecutor } from '../PipelineExecutor';
@@ -140,6 +142,7 @@ export function createPipelineExecutor(options: CreatePipelineExecutorOptions): 
     const pipelineExecutor: PipelineExecutor = (inputParameters: InputParameters): ExecutionTask =>
         createTask<PipelineExecutorResult>({
             taskType: 'EXECUTION',
+            title: pipeline.title,
             taskProcessCallback(updateOngoingResult: (newOngoingResult: PartialDeep<PipelineExecutorResult>) => void) {
                 return pipelineExecutorWithCallback(inputParameters, async (newOngoingResult) => {
                     updateOngoingResult(newOngoingResult);